Refactor InputProcessor into 2 distinct classes
