+++ src/enigma/FileInputProcessor.java
@@ -1,11 +1,12 @@
 /*
- * InputProcessor Class
+ * FileInputProcessor Class
  * Samuel Biondolillo
  * CIS220M:HY1 Object Oriented Programming
  * Goal: To handle the input and output routines for an Enigma emulator
  * Version 0.0.1 - 9/18/17
  *         0.0.2 - 9/22/17 - Renamed to InputProcessor, segregating input/output responsibilities
  *                           refactored code to reflect new responsibilities
+ *         0.0.3 - 10/20/17 - Refactored from InputProcessor into 2 classes: FileInputProcessor, KeyboardInputProcessor
  */
 
 package enigma;
@@ -15,34 +16,24 @@ import java.nio.file.*;
 import java.util.*;
 import interfaces.*;
 
-public class InputProcessor implements FileInput, KeyboardInput {
+public class FileInputProcessor implements FileInput {
 
 	private String messageIn;
-	private Scanner keyboardScanner;
 	private Scanner fileScanner;
-	private int inputMode;
 	
-	/*
-	 * Constructor
-	 * Default constructor uses keyboard input for processing
-	 */
-	public InputProcessor() {
-		keyboardScanner = new Scanner(System.in);
-		inputMode = 1;
-	}
+	
 	
 	/*
 	 * Constructor
 	 * @param filePath - String containing any valid file path
 	 */
-	public InputProcessor(String filePath) {
+	public FileInputProcessor(String filePath) {
 		try {
 			fileScanner = new Scanner(Paths.get(filePath));
 		}
 		catch (IOException e) {
 			System.out.println("Unable to access file: " + e);
 		}
-		inputMode = 0;
 	}
 	
 	/*
@@ -54,24 +45,6 @@ public class InputProcessor implements FileInput, KeyboardInput {
 	@Override
 	public Scanner getFileScanner() {
 		return fileScanner;
-	}
-	@Override
-	public Scanner getKeyboardScanner() {
-		return keyboardScanner;
-	}
-	public int getInputMode() {
-		return inputMode;
-	}
-	
-	/*
-	 * Reads in the next line from the keyboard and stores it in messageIn
-	 */
-	@Override
-	public void getKeyBoardIn() {
-		if (keyboardScanner.hasNextLine())
-			messageIn = keyboardScanner.nextLine();
-		else
-			messageIn = "";
 	}
 	
 	/*
